created RequestRideModal component to handle logic between modals
