+++ frontend/src/components/RequestRideModal/CustomRepeatingRides.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useCallback } from 'react';
+import React, { useEffect, useState } from 'react';
 import cn from 'classnames';
 import { useFormContext } from 'react-hook-form';
 import { Ride } from '../../types';
