+++ frontend/src/components/RequestRideModal/types.ts
@@ -1,0 +1,1 @@
+export type RideModalType = 'CREATE' | 'EDIT_REGULAR' | 'EDIT_SINGLE_RECURRING' | 'EDIT_ALL_RECURRING';
